first splitting of UnitTestFeatureGenerator
